compilation fix for gcc and clang

+++ include/universal/number/dbns/dbns_impl.hpp
@@ -704,7 +704,7 @@ protected:
 		assert(best_b >= 0); // second exponent is negative
 		int a = static_cast<unsigned>(-best_a);
 		int b = static_cast<unsigned>(best_b);
-		if (a < 0 || a > MAX_A || b > MAX_B) {
+		if (a < 0 || a > static_cast<int>(MAX_A) || b > static_cast<int>(MAX_B)) {
 			// try to project the value back into valid pairs
 			// the approximations of unity looks like (8,-5), (19,-12), (84,-53),... 
 			// they grow too fast and in a rather irregular manner. There are more
